+++ scm/driver/stash/org.go
@@ -18,6 +18,10 @@ func (s *organizationService) Find(ctx context.Context, name string) (*scm.Organ
 	return nil, nil, scm.ErrNotSupported
 }
 
+func (s *organizationService) FindMembership(ctx context.Context, name, username string) (*scm.Membership, *scm.Response, error) {
+	return nil, nil, scm.ErrNotSupported
+}
+
 func (s *organizationService) List(ctx context.Context, opts scm.ListOptions) ([]*scm.Organization, *scm.Response, error) {
 	return nil, nil, scm.ErrNotSupported
 }